Support selection multiple elements (#36)
* Support onMouseUp event on drawers

* Support multi element select

With this commit there are two things that are done. First this is that multiple elements can be selected holding down the CTRL button on the keyboard. The second change is that all selected elements are now dragged.

* Support snap lines with multi selection drag

* Implement lasso drawer

The lasso drawer will be shown when user is holding down a CTRL key and moving the mouse. Releasing CTRL key or the mouse will stop lasso drawing. The lasso drawing will also stop in case of any other mouse click.

* Support lasso elements selection

The lasso selection will use element bounding box in order to select elements.
All elements that are inside bounding box or intersects with bounding box will be selected.

* Support lasso theme